+++ src/globus/provision/chef-files/roles/domain-hadoop-slave.rb
@@ -1,0 +1,4 @@
+name "domain-hadoop-slave"
+description "A domain's Hadoop slave node"
+run_list "recipe[hadoop::hadoop-slave]"
+